refactor: changed contexts folder name to context.
